Refactor shop
