+++ tower-defense/src/shop/constants.rs
@@ -1,0 +1,16 @@
+use namui::*;
+
+pub const PADDING: Px = px(4.0);
+pub const SHOP_WH: Wh<Px> = Wh {
+    width: px(960.0),
+    height: px(480.0),
+};
+pub const SHOP_BUTTON_WH: Wh<Px> = Wh {
+    width: px(128.0),
+    height: px(48.0),
+};
+pub const SHOP_REFRESH_BUTTON_WH: Wh<Px> = Wh {
+    width: px(192.0),
+    height: px(48.0),
+};
+pub const SOLD_OUT_HEIGHT: Px = px(36.0);
