+++ tower-defense/src/shop/slot.rs
@@ -1,0 +1,12 @@
+use crate::game_state::item::Item;
+
+#[derive(Default, Clone)]
+pub enum ShopSlot {
+    #[default]
+    Locked,
+    Item {
+        item: Item,
+        cost: usize,
+        purchased: bool,
+    },
+}
